Improve ignore paths globifying
Ensures that if only a directory's basename is provided (without the preceeding path, just `dirname`), it prefixes `**/`

Currently when only a directory name is provided, it adds a postfix `/**` which is great for matching the entire tree ***if*** the tree is a *child* of the dir.  So `path` (converted to `path/**`) can filter paths like `path/to/file.js`, **but** if `path` is actually only a directory's basename it fails to match `/some/path/to/file.js`. This can be solved by prefixing `**/path`

Really helpful when only "node_modules" is provided as ignore path (like in `.gitignore` file) and it's actually some level deeper than current dir.

Also adjust the postfix to add just `**` instead of `/**` based on if last char is already a `/`